bug-in-delete-order-admin

+++ Shopping_Tutorial/Controllers/CartController.cs
@@ -27,7 +27,7 @@ namespace Shopping_Tutorial.Controllers
 			return View("~/Views/checkout/Index.cshtml");
 		}
 
-		public async Task<IActionResult> Add(int Id)
+		public async Task<IActionResult> Add(long Id)
 		{
 			ProductModel product = await _dataContext.Products.FindAsync(Id);
 			List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart") ?? new List<CartItemModel>();
